+++ sdk/storage/Azure.Storage.Files.Shares/tests/FileClientTests.cs
@@ -77,6 +77,29 @@ namespace Azure.Storage.Files.Shares.Tests
             Assert.AreEqual(createResponse.Value.ETag, propertiesResponse.Value.ETag);
         }
 
+        [Test]
+        public void Ctor_ConnectionString_CustomUri()
+        {
+            var accountName = "accountName";
+            var shareName = "shareName";
+            var directoryName = "directoryName";
+            var fileName = "fileName";
+            var filePath = $"{directoryName}/{fileName}";
+            var accountKey = Convert.ToBase64String(new byte[] { 0, 1, 2, 3, 4, 5 });
+
+            var credentials = new StorageSharedKeyCredential(accountName, accountKey);
+            var fileEndpoint = new Uri("http://customdomain/" + accountName);
+            var fileSecondaryEndpoint = new Uri("http://customdomain/" + accountName + "-secondary");
+
+            var connectionString = new StorageConnectionString(credentials, (default, default), (default, default), (default, default), (fileEndpoint, fileSecondaryEndpoint));
+
+            ShareFileClient fileClient = new ShareFileClient(connectionString.ToString(true), shareName, filePath);
+
+            Assert.AreEqual(accountName, fileClient.AccountName);
+            Assert.AreEqual(shareName, fileClient.ShareName);
+            Assert.AreEqual(filePath, fileClient.Path);
+        }
+
         [RecordedTest]
         public async Task Ctor_AzureSasCredential()
         {